final pass at restructuring codebase
