+++ internal/pkg/clipboard/clipboard.go
