+++ internal/pkg/clipboard/clipboard_test.go
