+++ internal/pkg/gif/gif.go
