+++ internal/pkg/gif/gif_test.go
